separate code execution logic from submit route into executeCode.js and hook up the evaluate button
