+++ server/api/routes/evaluate.js
@@ -1,12 +1,25 @@
 const express = require('express');
 const router = express.Router();
+const { models: { Test } } = require('../../db');
+const executeCode = require('./executeCode')
 
-router.post('/', async (req, res, next) => {
-  try {
-    console.log(req.body.code)
-    res.json('success!')
-  } catch(ex) {
-    next(ex);
+
+router.post('/:id', async (req, res, next) => {
+  // code is user code
+  let code = req.body.code
+
+      try {
+        let response = await Test.findOne({
+          where: { problemId: req.params.id }
+        });
+        
+        let problem = response.dataValues.test;
+
+        executeCode(code,problem,res)
+      
+    } catch(er){
+      console.log(er)
+      next(er)
   }
 });
